So there has been a lot of edits and definitely some unnecessary code has been created.  Using the copilot `simplify` option to clean stuff up.
Does seem like it doesnt remove unused methods, doing that manually though.  There was 6 methods hanging around that werent being used anymore.  Combined with the simplify, that reduceds the size of movie-vote.service.ts by like 2/3.

server.js - simpflying alone shortened to half the number of lines, shaved another 50 or so lines by asking it to remove any endpoints that werent being used in the app.  Though on testing found it got this wrong and topics were broken so just restored that and went to bed cause its late 🛌